inbound: Cache HTTP gateways per destination (#893)
When handling the HTTP gateway connections, the inbound proxy includes
the client's ID in the gateway's cache key so that a gateway service is
created for each (client, target) pair. Furthermore, the HTTP server
cache includes client _addresses_ in its key, so that a service is
cached for each _connection_ (which makes no sense).

In preparation for TCP gateway support, this change simplifies
inbound gateway handling by creating distinct target types used only by
the direct stack. Connections that lack identity or a target are failed
as early as possible, simplifying the gateway implementation. Client
identity is provided to gateway by setting the identity as a request
extension before the service is cached.