inbound: Use ALPN to determine transport header (#895)
The inbound proxy does not need to transparently detect transport
headers. The header is always indicated by TLS protocol negotiation.

This change decouples the `transport-header` module from the `Detect`
trait. The inbound direct stack is modified to require a transport
header when ALPN has determined that a header should be expected. Legacy
HTTP gateway support is attempted when no header is negotiated.

This sets up changes to the `Detect` server implementation.