detect: Return unknown protocol on detection timeout (#896)
The inbound proxy does not need to transparently detect transport
headers. The header is always indicated by TLS protocol negotiation.

This change decouples the `transport-header` module from the `Detect`
trait. The inbound direct stack is modified to require a transport
header when ALPN has determined that a header should be expected. Legacy
HTTP gateway support is attempted when no header is negotiated.

This sets up changes to the `Detect` server implementation.

* detect: Return unknown protocol on detection timeout

When performing HTTP detection, we always instrument a detection
timeout; and, if the timeout expires, we fail the connection. This leads
to confusion and complexity, especially when getting started with
Linkerd.

This change modifies the detect module to handle these timeouts
gracefully: instead of failing the connection with an error, we indicate
that no protocol could be detected and continue processing the
connection as a non-HTTP protocol. Furthermore, this change unifies the
timeout logic into the `DetectService` implementation to avoid
unnecessary boxing (due to async-trait).