Simplify opencensus stack composition (#900)
The tracing infrastructure exposes its implementation details (via type
signatures) to stack layers.

This change renames `app::core::spans` to `app::core::http_tracing`
which provides helpers that bundles the `TraceContext` layer with a
`SpanConverter`. It also provides a `OpenCensusSink` type alias so
that proxy stack builders can be ignorant of its actual type signature.

The `SpanSink` trait is updated to expose `is_enabled` which must be
called before constructing a span so that the `TraceContext` type isn't
required to hold an `Option<SpanSink>`.