Simplify outbound stack composition with builders (#901)
Outbound stack construction is instrumented from a mess of
free-functions that take a large number of arguments. Many stack
functions accept the same arguments. This leads to a lot of boilerplate
and makes it inconvenient to reconstruct stacks, as many values are
needed.

This change simplifies outbound stack construction by introducing two
new types:

- `core::ProxyRuntime`, which holds all common runtime configuration
  (metrics, tap, tracing, drain, etc); and
- `outbound::Outbound`, a helper that holds the outbound config, the
  proxy runtime, along with outbound-specific stack constructors.

All outbound stack constructors are now implemented in terms of these
types. This makes it simple to construct stacks without passing around
the same set of values ad-nauseam.