tracing: update to 0.1.23, remove `tracing-futures`
This branch updates the proxy's dependency on the `tracing` crate to
v0.1.23. It also removes the dependency on the `tracing-futures` crate
entirely, as the `Instrument` trait and `Instrumented` macro for
`std::future`s are now provided by `tracing` itself, and we don't use
any of the other functionality provided by `tracing-futures`. This
simplifies our dependency tree a bit and will allow us to stop allowing
an outdated `pin_project` in our cargo-deny config.

Signed-off-by: Eliza Weisman <eliza@buoyant.io>