Improve stack builder ergonomics
Some stack builders were not converted to be functions on the `Inbound`
or `Outbound` types because they had no obvious underlying stack type.

This change modifies `Inbound` and `Outbound` to be built with a
unit-typed stack by default. They each now have a set of common utility
functions used to bind the unit-typed stack to a service-stack.

No functional changes.