gateway: Transport TCP connections (#909)
The proxy's gateway mode can only transport HTTP requests; but, with the
introduction of the transport header (informed by TLS ALPN), the gateway
now has enough context to route TCP connections through the gateway.

This PR modifies the gateway stack to handle transported connections
differently from legacy HTTP connections. When clients connect to a
gateway with a transport header, the header's metadata is used for all
routing/discovery so, in the HTTP case, a request router is not
necessary. A TCP gateway is also introduced when a transport header does
not include a protocol hint. The same discovery is performed and
connections are dispatched through the TCP logical stack to balance and
forward connections within the gateway cluster.

More testing is needed on this, but I plan to address this in followup
changes that make the outbound stack more flexible and amenable to
testing.