Decouple application from TCP client
All clients--the inbound and outbound proxies, discovery clients,
etc--initiate TCP connections. This makes testing cumbersome, as the
application is tightly coupled to OS functionality.

This change makes the entire `App` generic over its connection strategy.
A `Connect` trait--a convenience over `Service`--is introduced so that
all of the relevant modules can be generic over the connector.