transport: Introduce address new-types
Throughout the codebase, we handle many types of `SocketAddr`s. It's
easy to use these instances in an incorrect context.

In preparation for upcoming changes to listen module and target types,
this change introduces a family of new-types to disambiguate these uses:

- `ClientAddr`  -- a client's (ephemeral) address
- `ServerAddr`  -- a server's address
- `OrigDstAddr` -- an address set by the `SO_ORIGINAL_DST` socket option

Additionally, `Local` and `Remote` address types are introduced to
further disambiguate the above types:

- `Remote<ClientAddr>` -- a server connection's peer address
- `Local<ServerAddr>`  -- a server connection's local address
- `Remote<ServerAddr>` -- a client connection's peer address
- `Local<ClientAddr>`  -- a client connection's local address

A special `ListenAddr` type is introduced to indicate an unbound address
to be used to bind a server. The `OrigDstAddr` trait is renamed to
`GetOrigDstAddr`.

No functional change.