make proxy builds work on windows
Currently, The proxy builds fail on windows because of the absence
of `future::preludes`. That does not seem to be required anymore
and the windows ctrl_c signal can be used similar with that of
the unix signals.

This PR updates the `shutdown` fn in imp for windows to be just
like the unix ones. I'm aware that the proxy is not intended to be ran
on windows machines but this was a easy fix to make builds work.

Signed-off-by: Tarun Pothulapati <tarunpothulapati@outlook.com>