detect: Surface timeouts to inner stack (#929)
In 95498395, the detect service was modified to handle detect timeouts
gracefully; but, we can't differentiate this situration from other
detection mismatches.

In order to use the detect trait more generally--especially for TLS
detection, this change modifies the `Detect` trait to pass an error into
the inner stack when detection fails. Uses are updated to either fail on
or log when an error is countered, as appropriate. This will be relied
on in followup changes.

Furthermore, the `Detect` trait's I/O type is moved to be a
trait-generic (rather than a method-generic) so that implementations may
impose further type constraints.