Remove the tcp_connection_duration_ms histogram
The TCP connection histogram is pretty useless in its current form: its
granularity in milliseconds is generally unsuited to long-lived
connections; and it ends up being extremely high cardinality.  And on
top of all that, the value isn't currently consumed in our UIs.

This change removes the `tcp_connection_duration_ms` histogram from the
proxy.

Relates to linkerd/linkerd2#5818