outbound: Decouple the resolver from concrete target types
The endpoint resolver module currently has a fixed dependency on the
`Concrete` and `Endpoint` target types.

This change extracts endpoint-type construction so that the HTTP logical
stack is agnostic of its endpoint target type. It also updates the
`EndpointFromMetadata` type to be responsible for marking identity as
disabled when appropriate. Finally, we are able to re-enable the
outbound TLS hinting test.