http: Parameterize header module
The module responsible for adding HTTP headers from target types expects
the target to be coerced to a `HeaderValue`. This means that a target
can only instrument a single header; and it creates some ambiguity.

This change updates the header insertion module to use a Param-type to
represent a key-value header pair, making uses clearer.