duplex: include src/dst peer in logs (#937)
The `linkerd_duplex` crate emits trace events for every read and write
operation performed, but these events don't make it clear which peer
performed the read or write. This branch adds names to the trace event
to make it possible to determine whether a read/write was performed by
the source or destination peer.

This does make the `Duplex` struct two words larger. Not sure if we
care. It might theoretically be possible to encode this using type
parameters to avoid the `&'static str`s; I'm not sure if this is worth
adding some wacky code to avoid two words of string constant pointer.

Signed-off-by: Eliza Weisman <eliza@buoyant.io>