outbound: Mock the original destination address in ingress-mode
The ingress-mode outbound proxy currently does not work as expected:
resolutions are performed for each unique endpoint so that a service
with 10 endpoints would manifest as 10 distinct balancers (with 20
resolutions).

While we should ultimately rework the outbound stack to have target
types that prevent this problem (i.e. by removing the original
destination address from the Logical target type), this change
implements a hack to avoid the observed cardinality issues: In
ingress-mode, the original destination address is now mocked out with an
IP address of 0.0.0.0 so that only a single cache entry is created per
logical destination.