Handle FailedPrecondition errors from the control plane
The control plane may return a FailedPrecondition error when the cluster
is in an unexpected state -- when there are more than one pod for a
single IP address. The proxy currently retries these lookups ad nauseum.

This change updates the proxies error handling to avoid retries when
this error is returned.