outbound: Configure endpoint construction in logical stack
ebf67d6b moved endpoint construction logic out of the logical stacks. In
doing so, the gateway was incorrectly configured to mark endpoints with
disabled identity, preventing mTLS between the gateway and destination
service.

This change moves endpoint construction into each logical stack so that
it need not be configured on each instantiation.