Update profile response to include a logical address (#954)
Profile responses optionally include a service name, but we only ever
use it in the context of an address (i.e. with a port).

This change introduces a `profiles::LookupAddr` param type that is used
as input into the profile client -- the lookup address may hold either a
named or numbered address. Service profile responses now include a
`LogicalAddr` that must be a named address.

This sets up for requiring that concrete addresses are named.