tls: Test TLS detection
7d2bdbb fixed an issue with TLS detection. This change adds a test for
this behavior.

In order to implement a test, this change modifies the `detect` helper
to not be responsible for SNI matching & TLS termination. The `detect`
helper is replaced with a `detect_sni` function that only reads the
ClientHello without terminating TLS. A test has been added for the
`detect_sni` helper.