admin: Handle connections that fail protocol detection
The HTTP detect module may fail to detect an HTTP protocol if the first
read does not return at least 14 bytes. This has caused spurious
failures as described in linkerd/linkerd2#5672.

This change updates the admin server's protocol detection to handle
these connections as HTTP/1 so that the HTTP server can try to process
them and fail on its own if the connection really is not HTTP.