Decouple server config from OrigDstAddr
Making the config types generic over the orig-dst strategy promulgates
generics across all of the stacks. But the binding logic is only
necessary at the very "top" of the stack where the accept loop is
driven.

This change eliminates the `GetAddrs` trait and modifies the `Bind`
trait to take params that allow the binding implementation to be
configured (i.e., from the environment). `OrigDstAddr` binding is
provided as a `Bind` implementation that wraps the `BindTcp` strategy.

This lets us eliminate the `mock-orig-dst` feature by moving the
`MockOrigDst` helpers into the integration test suite.