Initial fuzzer integration (#961)
Adds a set of fuzzers for various `linkerd` crates. 

The goal of this PR is to create an initial fuzzer infrastructure as well as
fuzzers that can be integrated with OSS-Fuzz for continuous fuzzing. 

The way I set it up was to place the fuzzers for a given crate within the
crate folder itself. I then put in a `fuzz_logic` module in the source code
that is to be fuzzed, thus following the low of how the existing tests are
written. The majority of additions in this PR are independent of the
linkerd2-proxy source code, but there are a few changes.

Signed-off-by: davkor <david@adalogics.com>