Fix caching in ingress mode (#965)
When the proxy is in `ingress` mode, requests with different
`l5d-dst-override` headers but the same address share the same key in the
cache. This causes consecutive requests to the same address to all go to the
value of the first `l5d-dst-override` header, ignoring the header values of
later requests—unless the services are evicted from the cache.

This occurs because each key into the cache is calculated based off
`LogicalAddr`'s `orig_dst` and `protocol` fields. When the proxy is in
`ingress` mode, the value of `orig_dst` is overridden to `0.0.0.0:port` This
results in an easy key conflict: two different requests with the same port
and protocol with result in the same key.

To fix this, the `logical_addr` field is added to `LogicalAddr` and is used
as a third field in calculating the cache key. `logical_addr` is the value of
the `l5d-dst-override` header which means requests to the same address but
with different header values will result in different keys.

Fixes linkerd/linkerd2#5975

Signed-off-by: Kevin Leimkuhler <kevin@kleimkuhler.com>