discover: replace `linkerd-channel` with `tokio-util` `PollSender` (#969)
This commit replaces the use of `linkerd-channel`'s pollable MPSC in the
`linkerd_discover` crate's `buffer` module with `tokio-util`'s
[`sync::PollSender`][1] type, which also implements a pollable sender.
The `PollSender` type is likely more efficient than our channel
implementation, since it uses the [`ReusableBoxFuture`][2] type from
`tokio-util`, which avoids allocating a `Box` for _every_ message sent
to the channel, and instead reuses a single allocation. It also wraps
the `tokio::sync::mpsc` `Sender`, rather than reimplementing the
capacity limit on top of the `UnboundedSender` type. This reduces
complexity and should hopefully be less error-prone.

[1]: https://docs.rs/tokio-util/0.6.5/tokio_util/sync/struct.PollSender.html
[2]: https://docs.rs/tokio-util/0.6.5/tokio_util/sync/struct.ReusableBoxFuture.html

Signed-off-by: Eliza Weisman <eliza@buoyant.io>