outbound: skip logical stacks when no profile is discovered (#963)
Currently, the outbound proxy's `Logical` target type contains an
`Option<profiles::Receiver>`. When no profile is discovered, this field
is set to `None`; the per-profile layers in the logical stack do nothing
when this field is `None`. However, we still *construct* this whole
stack for all destinations, even those without logical addresses.

This PR refactors the outbound proxy so that the `Logical` targets
*always* have discovered profiles, and the logical stack is skipped when
no profile is discovered. Profile discovery still returns an
`Option<profiles::Receiver>`, but we unwrap it in one place immediately
after the profile discovery layer, and skip directly to the endpoint
stack when it is `None`.

This means that we can now avoid constructing layers like traffic split
etc when there is no profile for a target, rather than building a
traffic split that does nothing.

Note that some of the layers in the logical stacks currently *do* still
have code for handling the case where a `Logical` target has no profile.
These layers currently take targets that are bounded with
`Param<Option<profiles::Receiver>>`. The `Logical` type still has its
`Param<Option<profiles::Receiver>>` impl, it just always returns `Some`
now. I chose to leave this code as-is since this branch is already a
fairly large refactor. We can simplify a lot of this code in a follow-up
(although some layers may also require a similar change on the inbound
side).

Signed-off-by: Eliza Weisman <eliza@buoyant.io>