admin: Assume meshed connections are HTTP/2
We've recently updated the admin server to assume that protocol
detection timeouts indicate that the client was just HTTP/1; but this
will almost never be true for meshed connections, which are transported
over HTTP/2.

This change updates the fallback logic to account for this. Detection
timeouts on meshed connections are now assumed to be HTTP/2. When we get
TLS connections that we can't terminate, we now return an error message
that includes the SNI value to more clearly indicate that the client
expected an alternate server identity.