fuzz: add logging to fuzz targets
In order to make debugging fuzz test failures easier, it's useful to be
able to get logs from the proxy and libraries during a fuzz run.
However, because running the fuzz targets with a fuzzer will generate a
massive number of imputs, we probably don't want verbose logs by default
--- this would produce a huge amount of mostly useless data.

This branch adds a simple tracing setup to the fuzz targets. When
running without the `RUST_LOG` environment variable set, each run won't
log anything. This means the current behavior on cluster-fuzz is
unchanged and we won't output giant amounts of logs. However, when
running a minimized reproducer locally for debugging, we can enable any
amount of logging by setting `RUST_LOG`, which should make it easier to
diagnose fuzz failures.