fuzz: allow client requests to fail  (#989)
Currently, the inbound HTTP fuzz tests are failing when the fuzzer
generates invalid HTTP headers and the proxy rejects the request by
closing the connection.

This is because the test-support HTTP request helper expects the client
request to never fail --- valid behavior for the tests where this is
used, but not for the fuzz logic. This branch changes these functions to
return `Result`s so that the test that uses them can choose whether or
not to panic.

The test helper code currently uses `expect` messages to provide context
for _where_ a particular error occurred (e.g. were we driving the client
to readiness, or actually making a request?). Changing the test helpers
to return `Result`s would remove some of this context. Therefore, I
added a quick error wrapper type to wrap an error with a message,
allowing us to indicate where an error occurred while still returning a
`Result` that the test can choose to unwrap.

Signed-off-by: Eliza Weisman <eliza@buoyant.io>