ingress: Require the l5d-dst-override header
The outbound ingress-mode proxy generally relies on the
`l5d-dst-override` header to be set, but it still tries to handle other
cases (mostly to satisfy type constraints we needed previously). This is
unnecessary.

This change modifies the ingress-mode outbound proxy to require that the
`l5d-dst-override` header be set to a named address (i.e., not an IP
address). When this header isn't set, requests are failed a descriptive
error is emitted.

Furthermore, the inbound proxy no longer honors the `l5d-dst-override`
header. It never should have supported this header, but this was likely
copied from the outbound router. The header is only intended for
ingresses.