http: Remove `identity_from_header` helper
The `proxy::http::identity_from_header` helper is only used in one place
and doesn't need to be exposed by the proxy::http module. Similarly the
`L5D_REQUIRE_ID` constant does not need to be used outside of the
outbound HTTP endpoint stack, though it is defined in `app::core`.

This change renames the `require_identity_on_endpoint` to
`require_id_header` and consolidates all related logic & constants into
the module. The module is now responsible for clearing the header as
well (rather than pairing the module with a generic strip-header layer).