inbound: Set the l5d-client-id header (#981)
Applications have no way to know whether a request was transported
via Linkerd's mTLS and, if it was, what the client's identity was.

This change sets the `l5d-client-id` header with the value of the client's
identity name. Furthermore, any prior values for this header are stripped
from inbound requests so that client identities may not be spoofed.

This header is set on all inbound requests over meshed connections. This
behavior may be made configurable in the future. 

Related to linkerd/linkerd2#3342 and linkerd/linkerd2#3756